Hiding assets/activities & assessments for new-style courses. Hiding 'Revert to Javascript assessment' button for new-style assessments. Changing lesson weight validator from int to float. 	Change on 2014/09/23 by borislavr <borislavr@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=76172008

+++ coursebuilder/modules/dashboard/unit_lesson_editor.py
@@ -398,7 +398,7 @@ class UnitTools(object):
 
         self._apply_updates_common(unit, entity_dict, errors)
         try:
-            unit.weight = int(entity_dict.get('weight'))
+            unit.weight = float(entity_dict.get('weight'))
             if unit.weight < 0:
                 errors.append('The weight must be a non-negative integer.')
         except ValueError:
